Improve translation for Chinese language in Settings
The changes were made to improve the translation quality of some phrases and words in the settings of the program for the Chinese language. Certain phrases made better sense when replaced with different words, providing more context and clarity to the user. It is important that the quality of the translation is maintained to facilitate Chinese-speaking